python: Fix buggy call to PyArg_Parse due to type of len
PyArg_Parse expect the type of the len param to be Py_ssize_t but
instead it was passed an int. This did not lead to any issues on the mac
but on linux it caused a segfault.

This commit fixes that issue and also does some housekeeping.

+++ python/osdp_sys/utils.c
@@ -232,7 +232,7 @@ int pyosdp_dict_get_object(PyObject *dict, const char *key, PyObject **obj)
 
 static int channel_read_callback(void *data, uint8_t *buf, int maxlen)
 {
-	int len;
+	Py_ssize_t len;
 	PyObject *channel = data;
 	uint8_t *tmp;
